% Generated by roxygen2 (4.0.2): do not edit by hand
\name{binned_function}
\alias{binned_function}
\title{calculate binned values}
\usage{
binned_function(bins, numvar, binfun = "mean", mcolname = "avg")
}
\arguments{
\item{bins}{a \code{GRanges} object representing genomic bins}

\item{numvar}{a named \code{RleList} object representing numerical variable along the genome}

\item{binfun}{character string of the function to apply}

\item{mcolname}{the name of the metadata column containing the binned value of the object}
}
\value{
bins with the \code{mcolname} containing the binned value
}
\description{
given a set of genomic bins and an RLE-list object, return value
}
\details{
Adapted from documentation of \code{tileGenome}
}

