% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bootstrap_correlation}
\alias{bootstrap_correlation}
\title{run bootstrapped correlation}
\usage{
bootstrap_correlation(data, data_columns, log_transform = TRUE,
  non_zero = "either", n_boot = 1000)
}
\arguments{
\item{data}{the data we are working with}

\item{data_columns}{which columns to use}

\item{log_transform}{do a log transformation on the data before calculating the correlation}

\item{non_zero}{which columns can be non_zero}

\item{n_boot}{how many bootstrap samples to generate}
}
\value{
correlation value and standard deviation
}
\description{
given a set of points for correlations, generate bootstrapped samples and calculate a 95%CI or SD
}

