% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bootstrap_odregress}
\alias{bootstrap_odregress}
\title{boostrap orthogonal regression}
\usage{
bootstrap_odregress(x_data, y_data, n_boot = 500)
}
\arguments{
\item{x_data}{the X-data}

\item{y_data}{the Y-data}

\item{n_boot}{how many bootstrap samples to do}
}
\value{
vector of the coefficients
}
\description{
for a set of data, do a bootstrapped orthogonal regression
}

