% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert_to_counted}
\alias{convert_to_counted}
\title{convert to counted GRanges}
\usage{
convert_to_counted(reads_file, delim = ",", read_start = "startx",
  width = 1, strand = "strand", scramble_strand = TRUE, max_count = 6,
  prepend = "")
}
\arguments{
\item{reads_file}{the file with the reads}

\item{delim}{the delimiter to use}

\item{read_start}{the column containing the read starts}

\item{width}{how wide are the reads}

\item{strand}{which column name has the strand information}

\item{scramble_strand}{should the strand be scrambled (default is \code{TRUE})}

\item{max_count}{what is the maximum value that the counts should be}

\item{prepend}{text to prepend the out_file name with}
}
\value{
the RData filename
}
\description{
given a set of reads in a delimited file, find the unique locations, count the number of reads at each location,
apply a hard threshold if required, and save the data to an RData file.
}

