% Generated by roxygen2 (4.0.2): do not edit by hand
\name{correlate_non_zero}
\alias{correlate_non_zero}
\title{correlate non-zeros}
\usage{
correlate_non_zero(data, data_columns, log_transform = TRUE,
  non_zero = "either", test = TRUE)
}
\arguments{
\item{data}{the data we are working with}

\item{data_columns}{which columns to use}

\item{log_transform}{do a log transformation on the data before calculating the correlation}

\item{non_zero}{"either" or "both" values need to be non-zero}

\item{test}{also return a p-value of the correlation?}
}
\value{
numeric vector
}
\description{
from a \code{DataFrame} object, generate correlation of the non-zero entries
}

