% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cum_indices}
\alias{cum_indices}
\title{cumulative indices}
\usage{
cum_indices(indices_list, direction = "low_to_high")
}
\arguments{
\item{indices_list}{the list of indices we want to work with}

\item{direction}{which way to put the indices together (low_to_high or high_to_low)}
}
\value{
new indices_list with cumulative indices
}
\description{
Given a list of indices assumed to in ranked order, generate a set of cumulative
indices wherein the indices are essentially collapsed together.
}
\examples{
indices_list <- list("1" = c(1,2,3), "2" = c(4,5,6), "3" = c(7,8,9))
cum_indices(indices_list, "low_to_high")
cum_indices(indices_list, "high_to_low")
}

