% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cut_2_value}
\alias{cut_2_value}
\title{cut names 2 value}
\usage{
cut_2_value(cuts, type = "left")
}
\arguments{
\item{cuts}{the set of character \code{cuts} to process}

\item{type}{which type of value to return}
}
\value{
numerical vector
}
\description{
Given a set of \code{cuts} as character names of a vector, return either the \code{left},
\code{right} or \code{mid} as a numerical value that can be used for plotting.
}
\examples{
# set up some fake data
cuts <- c("[0.301,0.4771)", "[0.4771,0.6276)")

cut_2_value(cuts, "left")
cut_2_value(cuts, "right")
cut_2_value(cuts, "mid")
}

