% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate_quantile_indices}
\alias{generate_quantile_indices}
\title{quantile index}
\usage{
generate_quantile_indices(x, n_quantile = 101)
}
\arguments{
\item{x}{the data to generate quantiles for}

\item{n_quantile}{how many quantiles to split into}
}
\value{
list with indices for each quantile
}
\description{
Given a set of values, generate the indices of the data for supplied quantiles
}

