% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_overlap_counts}
\alias{get_overlap_counts}
\title{get overlap counts}
\usage{
get_overlap_counts(file_list, granges, offset = 0)
}
\arguments{
\item{file_list}{the list of files to read in corresponding to ranges}

\item{granges}{the genomic ranges to count the overlaps with}

\item{offset}{a value by which to \code{shift} the reads by, positively for "+" and negatively for "-"}
}
\value{
vector of counts
}
\description{
given a set of files (assumed to be chromosomes), and a set of \code{GRanges}, count the number of overlaps with
the \code{GRanges}.
}

