% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list_correlation}
\alias{list_correlation}
\title{list correlation}
\usage{
list_correlation(x, indices_list, similarity = stats::cor)
}
\arguments{
\item{x}{the data.frame of data (should be two columns only)}

\item{indices_list}{the list of indices to be iterated over}

\item{similarity}{the function used to calculate the similarity}
}
\value{
numerical vector of correlations
}
\description{
Given a data.frame of data, and a list of indices, calculate the correlation
between the columns of the data.frame using the indices, iterating over the list
of indices.
}

