% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rle_mean_nozero}
\alias{rle_mean_nozero}
\title{calculate Rle mean without zeros}
\usage{
rle_mean_nozero(in_rle)
}
\arguments{
\item{in_rle}{the \code{Rle} to calculate the non-zero mean for}
}
\value{
a numeric value
}
\description{
Given an \code{Rle} with possible zeros, first remove the zeros, and then calculate the weighted mean
}

