% Generated by roxygen2 (4.0.2): do not edit by hand
\name{run_cum_quantiles}
\alias{run_cum_quantiles}
\title{run adding quantiles}
\usage{
run_cum_quantiles(x, n_quantile = 101, similarity = stats::cor,
  cut_loc = "left")
}
\arguments{
\item{x}{a two column data.frame}

\item{n_quantile}{how many quantiles}

\item{cut_loc}{which way to generate cut values on the data ("left", "right", "mid")}

\item{similarity_function}{what similarity function to use}
}
\value{
data.frame with the values, the corresponding cut value, and which cumulative indices generated it
}
\description{
Given a data set, generate the quantiles, the \code{cum_indices}, run the correlation,
and return a data.frame for plotting and diagnostics.
}

